Avoid "You tried to plan twice” message
The skip condition throws a "You tried to plan twice" message in some cases (cf. CPANTesters reports). 
The duplicate test for the lib was removed and the planning was moved behind the skip tests condition. Should fix it.

cf. https://stackoverflow.com/questions/21626659/how-to-avoid-you-tried-to-plan-twice-at-testmore + https://github.com/hoytech/Log-File-Rolling/issues/1